MISUV-7388:   * refactor ClaimToAdjustService
